+++ LBEngine/System.cpp
@@ -12,6 +12,6 @@ void System::AddEntity(Entity* pEntity)
 
 void System::SubscribeToComponentType(ComponentType type)
 {
-	std::bitset<COMPONENT_TYPE_COUNT> add = 1 << (size_t)type;
+	std::bitset<COMPONENT_TYPE_COUNT> add = (size_t)1 << (size_t)type;
 	m_componentsMask |= add;
 }
